Conflicts:
	src/main/java/com/amee/platform/resource/dataitem/v_3_0/DataItemBuilder_3_0_0.java
	src/main/java/com/amee/platform/resource/dataitem/v_3_0/DataItemsBuilder_3_0_0.java
	src/main/java/com/amee/platform/resource/itemdefinition/v_3_1/ItemDefinitionBuilder_3_1_0.java
	src/main/java/com/amee/platform/resource/itemvaluedefinition/v_3_0/ItemValueDefinitionBuilder_3_0_0.java
	src/main/java/com/amee/platform/resource/itemvaluedefinition/v_3_1/ItemValueDefinitionsBuilder_3_1_0.java
	src/main/java/com/amee/platform/resource/returnvaluedefinition/v_3_1/ReturnValueDefinitionBuilder_3_1_0.java
	src/main/java/com/amee/platform/resource/returnvaluedefinition/v_3_1/ReturnValueDefinitionRemover_3_1_0.java
	src/main/java/com/amee/platform/resource/returnvaluedefinition/v_3_1/ReturnValueDefinitionsBuilder_3_1_0.java

+++ src/main/java/com/amee/platform/resource/returnvaluedefinition/v_3_1/ReturnValueDefinitionRemover_3_1_0.java
@@ -1,10 +1,14 @@
 package com.amee.platform.resource.returnvaluedefinition.v_3_1;
 
 import com.amee.base.domain.Since;
-import com.amee.base.resource.*;
+import com.amee.base.resource.MissingAttributeException;
+import com.amee.base.resource.NotFoundException;
+import com.amee.base.resource.RequestWrapper;
+import com.amee.base.resource.ResponseHelper;
 import com.amee.domain.data.ItemDefinition;
 import com.amee.domain.data.ReturnValueDefinition;
 import com.amee.platform.resource.returnvaluedefinition.ReturnValueDefinitionResource;
+import com.amee.service.auth.ResourceAuthorizationService;
 import com.amee.service.definition.DefinitionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
@@ -18,6 +22,9 @@ public class ReturnValueDefinitionRemover_3_1_0 implements ReturnValueDefinition
     @Autowired
     private DefinitionService definitionService;
 
+    @Autowired
+    private ResourceAuthorizationService resourceAuthorizationService;
+
     public Object handle(RequestWrapper requestWrapper) {
         String itemDefinitionIdentifier = requestWrapper.getAttributes().get("itemDefinitionIdentifier");
         if (itemDefinitionIdentifier != null) {
@@ -27,6 +34,10 @@ public class ReturnValueDefinitionRemover_3_1_0 implements ReturnValueDefinition
                 if (returnValueDefinitionIdentifier != null) {
                     ReturnValueDefinition returnValueDefinition = definitionService.getReturnValueDefinitionByUid(itemDefinition, returnValueDefinitionIdentifier);
                     if (returnValueDefinition != null) {
+                        // Authorized?
+                        resourceAuthorizationService.ensureAuthorizedForRemove(
+                                requestWrapper.getAttributes().get("activeUserUid"), returnValueDefinition);
+                        // Handle ReturnValueDefinition removal.
                         definitionService.remove(returnValueDefinition);
                         definitionService.invalidate(returnValueDefinition.getItemDefinition());
                         return ResponseHelper.getOK(requestWrapper);
